+++ src/axom/quest/examples/quest_marching_cubes_example.cpp
@@ -1611,10 +1611,9 @@ int allocatorIdToTest(axom::runtime_policy::Policy policy)
   //---------------------------------------------------------------------------
   // Memory resource.  For testing, choose device memory if appropriate.
   //---------------------------------------------------------------------------
-  int allocatorID =
-    policy == RuntimePolicy::seq ? axom::detail::getAllocatorID<axom::MemorySpace::Host>() :
+  int allocatorID = policy == RuntimePolicy::seq ? axom::detail::getDefaultHostAllocatorID() :
   #if defined(AXOM_RUNTIME_POLICY_USE_OPENMP)
-    policy == RuntimePolicy::omp ? axom::detail::getAllocatorID<axom::MemorySpace::Host>()
+    policy == RuntimePolicy::omp ? axom::detail::getDefaultHostAllocatorID()
     :
   #endif
   #if defined(AXOM_RUNTIME_POLICY_USE_CUDA)